basic Hebrew
Please note I wanted to push something rather than nothing so its partial for now.
I've used a female form as suggested, when referring the user or other users. I used the male form when referring to objects (assuming the object is a male) but it might be wrong. I.e. a Message is a female while a Button is a male.
There might be some mismatches or wrong translation. Only once someone look at the translated UI he can and figure out the context of the word. I guess it will be ongoing with others to continue and translate as they use it, not sure when and if I'll continue the work here...